requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
redis==5.0.1
python-dotenv==1.0.0
httpx==0.26.0
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
openai==1.10.0
anthropic==0.18.0
pillow==10.2.0
python-multipart==0.0.6
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
tenacity==8.2.3
requests==2.31.0
